asdf

+++ ft_strjoin.c
@@ -6,12 +6,13 @@
 /*   By: eelisaro <eelisaro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/29 21:33:17 by eelisaro          #+#    #+#             */
-/*   Updated: 2022/10/29 21:49:52 by eelisaro         ###   ########.fr       */
+/*   Updated: 2022/11/08 17:00:34 by eelisaro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include <stdlib.h>
+#include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
@@ -28,8 +29,8 @@ size_t	ft_strlen(const char *s)
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*nstr;
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	j = 0;
 	i = 0;
@@ -52,3 +53,31 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	nstr[i] = '\0';
 	return (nstr);
 }
+// char	*ft_strjoin(char const *s1, char const *s2)
+// {
+// 	int		i;
+// 	int		len1;
+// 	int		len2;
+// 	char	*str;
+
+// 	if (s1 && s2)
+// 	{
+// 		len1 = ft_strlen(s1);
+// 		len2 = ft_strlen(s2);
+// 		str = (char*)malloc(sizeof(char) * (len1 + len2 + 1));
+// 		if (str == NULL)
+// 			return (NULL);
+// 		i = -1;
+// 		while (s1[++i])
+// 			str[i] = s1[i];
+// 		i = -1;
+// 		while (s2[++i])
+// 		{
+// 			str[len1] = s2[i];
+// 			len1++;
+// 		}
+// 		str[len1] = '\0';
+// 		return (str);
+// 	}
+// 	return (NULL);
+// }
